# feat: improve Decoder string unescaping performance