… internal/decoder/string.go